Model/budget-grocery-app/requirements.txt
pandas
scikit-learn
lightgbm
sentence-transformers
faiss-cpu
fastapi
uvicorn
pydantic
mlflow
pytest